Split monolithic module into smaller pieces
