+++ src/lib/Burrito/Type/Template.hs
@@ -1,0 +1,11 @@
+module Burrito.Type.Template
+  ( Template(..)
+  ) where
+
+import qualified Burrito.Type.Token as Token
+
+
+-- | Represents a URI template.
+newtype Template = Template
+  { tokens :: [Token.Token]
+  } deriving (Eq, Show)
